requirements.txt
appdirs==1.4.3
CacheControl==0.12.6
cachetools==4.2.1
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
contextlib2==0.6.0
Deprecated==1.2.12
distlib==0.3.0
distro==1.4.0
google-api-core==1.26.1
google-api-python-client==2.0.2
google-auth==1.28.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.3
googleapis-common-protos==1.53.0
html5lib==1.0.1
httplib2==0.19.0
idna==2.8
ipaddr==2.2.0
lockfile==0.12.2
msgpack==0.6.2
oauthlib==3.1.0
packaging==20.3
pep517==0.8.2
progress==1.5
protobuf==3.15.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyGithub==1.54.1
PyJWT==1.7.1
pyparsing==2.4.6
python-dotenv==0.15.0
pytoml==0.1.21
pytz==2021.1
requests==2.22.0
requests-oauthlib==1.3.0
retrying==1.3.3
rsa==4.7.2
six==1.14.0
uritemplate==3.0.1
urllib3==1.25.8
webencodings==0.5.1
wrapt==1.12.1
